Lock this library to versions of nocarrier/hal >=0.9.4 and <0.9.7
Since this library inherits from nocarrier/hal (instead of using
composition), changes to nocarrier/hal broke the functionality of this
library. This change is a patch fix to put a band-aid on the issue.
Future versions of this library will use composition and the decorator
pattern to wrap nocarrier/hal, rather than inherit from it.

This fixes #4.